Some newer files

+++ general-xorg-files/.fvwm/matchbox-desktop.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
 
-if ! pkill matchbox; then
-    matchbox-keyboard & disown
+if [ "$FVWM_CURRENTMODE" = 'desktop' ]; then
+    if ! pkill matchbox; then
+        echo "Exec exec matchbox-keyboard"
+    fi
+else
+    echo "All (Keyboard) Iconify"
 fi
